requirements.txt
py-cord
python-dotenv
flask
pymongo
